create generic search bar for Finance, Manage Appointments, and Estimates page
