+++ src/hooks/FinanceHooks.tsx
@@ -273,194 +273,3 @@ export function RenderClientFinance(props: TableContent){
         </table>
     )
 }
-
-interface Search{
-    searchValue:string, 
-    setSearchValue: (e:string)=>void, 
-    setAppointments: (e:ClientFinance[])=>void,
-    suggestions: React.JSX.Element | undefined,
-    setSuggestions: (e:React.JSX.Element)=>void,
-    hidden: boolean,
-    setHidden: (e:boolean) => void
-}
-
- export function SearchBar(props: Search){
-
-    function includeResults(appointments: ClientFinance[], appointmentFields: string[]){
-        const filteredAppointments = appointments.filter((appointment:any) => {
-            return Object.values(appointment)
-                .some(value => typeof value === "string" ? value.toLowerCase().includes(props.searchValue.toLowerCase()) : "");
-        });
-
-        const suggestedValues = appointmentFields.flatMap(field =>
-            filteredAppointments
-                .filter((appointment:any) => appointment[field].toLowerCase().includes(props.searchValue.toLowerCase()))
-        );
-
-        return suggestedValues;
-    }
-
-    function exactResults(appointments: ClientFinance[], appointmentFields: string[]){
-        const filteredAppointments = appointments.filter((appointment:any) => {
-            return Object.values(appointment)
-                .some(value => typeof value === "string" ? value.toLowerCase() === props.searchValue.toLowerCase() : "");
-        });
-
-        const suggestedValues = appointmentFields.flatMap(field =>
-            filteredAppointments
-                .filter((appointment:any) => appointment[field].toLowerCase() === props.searchValue.toLowerCase())
-        );
-
-        return suggestedValues;
-    }
-
-
- 
-    async function searchResults(){
-        try{
-            const finances = await api.listDocuments(process.env.REACT_APP_CART_DATABASE_ID, process.env.REACT_APP_FINANCE_PAYMENTS_COLLECTION_ID);
-
-            const exactValues = exactResults(finances.documents, ['email', 'financeTotal' ,'type'])
-            const includeValues = includeResults(finances.documents, ['email', 'financeTotal' ,'type'])
-
-
-           if(exactValues.length && exactValues){
-               props.setAppointments(exactValues)
-           }else if(includeResults.length && includeValues){
-               props.setAppointments(includeValues)
-           }else{
-               props.setAppointments([])
-           }
-        }catch(err){
-            console.error(err);
-        }
-    }
-
-    function AutoSuggest(searchValue: string){
-
-            useEffect(()=>{
-                async function searchSuggest(){
-                    try{
-                        const finances = await api.listDocuments(process.env.REACT_APP_CART_DATABASE_ID, process.env.REACT_APP_FINANCE_PAYMENTS_COLLECTION_ID);
-
-                        const filteredAppointments = finances.documents.filter((appointment:any) => {
-                            return Object.values(appointment)
-                                .some(value => typeof value === "string" ? value.toLowerCase().includes(props.searchValue.toLowerCase()) : "");
-                        });
-                
-                        const suggestedValues =   ['email', 'financeTotal', 'type'].flatMap(field =>
-                            filteredAppointments
-                                .filter((appointment:any) => appointment[field].toLowerCase().includes(props.searchValue.toLowerCase()))
-                                .map((appointment:any) => appointment[field])
-                        );
-
-                        const removeDuplicates:string[] = [];
-                        suggestedValues.forEach((value:string)=>removeDuplicates.indexOf(value) === -1 ? removeDuplicates.push(value) : "")
-
-
-                        let i = 0
-                        const list = removeDuplicates.map((value:string)=>{
-                            return(
-                                <option key = {i++}>{value}</option>
-                            )
-                        })
-
-                        const results = <select
-                        onChange = {(e)=>props.setSearchValue(e.target.value)}
-                        >{list}</select>
-                        
-
-                        props.setSuggestions(results);
-                    }catch(err){
-                        console.error(err);
-                    }
-                }
-
-                searchSuggest()
-            },[searchValue])
-
-            return props.suggestions;
-
-    }
-
-
-    //type search value
-    //as a new letter is entered in search bar, list possible choices user can make based on the letters in search bar so far
-    //like if the user inputs el, suggest "Kyrielight", "Elizabeth" or "Fuel Systems".
-    //if the user clicks on one of these suggestions, the search value becomes that suggestion 
-
- 
-
-    async function filterByValue(filter: string){
-      try{
-        const finances = await api.listDocuments(process.env.REACT_APP_CART_DATABASE_ID, process.env.REACT_APP_FINANCE_PAYMENTS_COLLECTION_ID);
-
-        if(filter === "email"){
-            const sortAppointments = finances.documents.sort((a:any, b: any)=>a.email.localeCompare(b.email));
-            props.setAppointments(sortAppointments);
-        }else if(filter === "type"){
-            const sortAppointments = finances.documents.sort((a:any, b: any)=>a.type.localeCompare(b.type));
-            props.setAppointments(sortAppointments);
-        }else if(filter === "financeTotal"){
-            const sortAppointments = finances.documents.sort((a:any, b: any)=>Number(a.financeTotal) - Number(b.financeTotal));
-            props.setAppointments(sortAppointments);
-        }
-      }catch(err){
-        console.error(err);
-      }
-    }
-
-    //filter buttons
-    //select from date, model, make, year or service, name
-    //when selected, it will re-render appointments based on that filter alphabetically/recent-latest
-
-
-    function searchFilters(){
-        return(
-            <section className = "filters flex justifyBetween alignCenter flex-col">
-
-                <i className = "fa-solid fa-xmark button" id = "button" onClick = {()=>props.setHidden(!props.hidden)}></i>
-
-                {Button({text: "Filter By Email", handleButtonClick: (e)=>{
-                    e.preventDefault();
-                    filterByValue("email");
-                }})}
-        
-                {Button({text: "Filter By Card Type", handleButtonClick: (e)=>{
-                    e.preventDefault();
-                    filterByValue("type");
-                }})}
-
-                {Button({text: "Filter By Finance Total", handleButtonClick: (e)=>{
-                    e.preventDefault();
-                    filterByValue("financeTotal");
-                }})}
-        
-        
-        </section>
-        )
-    }
-
-    return(
-        <form className = "flex justifyCenter">
-            <section className = "flex flex-col alignCenter search">
-
-            {props.hidden ? searchFilters() : ""}
-                <div className="flex alignCenter justifyEvenly">
-                    <input type = "search" value = {props.searchValue} onChange = {(e)=>props.setSearchValue(e.target.value)}/>          
-                    {AutoSuggest(props.searchValue)}
-                    {Button({text: "Search", handleButtonClick: (e)=>{
-                    e.preventDefault();
-                    searchResults()}})}
-                                <i className = {`${props.hidden ? "fa-solid fa-caret-up clearButton" : "fa-solid fa-caret-down clearButton"}`} onClick = {()=>props.setHidden(!props.hidden)}></i>
-
-
-                </div>
-
-            </section>
-
-    
-        </form>
-       
-    )
- }
